server/requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pillow==10.4.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.2.2
